Move all DMA specific operations into a new dma.py file. Import back into pull_data.py.
